chore(client): cleanup precompile acceleration using a macro
